requirements.txt
dj-database-url==0.4.1
Django==1.9.8
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
django-webpack-loader==0.4.1
djangorestframework==3.5.3
ipython
